refactor search bar logic into custom hook
