+++ webapp/src/app/utils/PlaceAutocomplete.jsx
@@ -1,0 +1,33 @@
+import { useEffect, useRef, useState } from "react";
+import { useMapsLibrary } from "@vis.gl/react-google-maps";
+
+const PlaceAutocomplete = ({ children, onPlaceSelect, onChange }) => {
+  const [placeAutocomplete, setPlaceAutocomplete] = useState(null);
+  const inputRef = useRef(null);
+  const places = useMapsLibrary("places");
+
+  useEffect(() => {
+    if (!places || !inputRef.current) return;
+
+    const options = {
+      fields: ["geometry", "name", "formatted_address"],
+    };
+
+    setPlaceAutocomplete(new places.Autocomplete(inputRef.current, options));
+  }, [places]);
+
+  useEffect(() => {
+    if (!placeAutocomplete) return;
+
+    placeAutocomplete.addListener("place_changed", () => {
+      onPlaceSelect(placeAutocomplete.getPlace());
+    });
+  }, [placeAutocomplete]);
+  return (
+    <div className="autocomplete-container">
+      <input ref={inputRef} onChange={onChange} value={children} />
+    </div>
+  );
+};
+
+export default PlaceAutocomplete;
